+++ crates/rpc/src/lib.rs
@@ -361,6 +361,8 @@ pub mod test_utils {
         let class1_hash = class_hash_bytes!(b"class 1 hash");
         let class2_hash = class_hash_bytes!(b"class 2 hash (sierra)");
         let class_hash_pending = class_hash_bytes!(b"class pending hash");
+        let sierra_class = SierraHash(class2_hash.0);
+        let sierra_casm_hash = casm_hash_bytes!(b"non-existent");
 
         let storage_addr = storage_address_bytes!(b"storage addr 0");
 
@@ -377,6 +379,7 @@ pub mod test_utils {
             );
 
         let state_update2 = StateUpdate::default()
+            .with_declared_sierra_class(sierra_class, sierra_casm_hash)
             .with_deployed_contract(contract2_addr, class2_hash)
             .with_contract_nonce(contract1_addr, contract_nonce!("0x10"))
             .with_contract_nonce(contract2_addr, contract_nonce!("0xfeed"))
@@ -400,25 +403,17 @@ pub mod test_utils {
         let sierra_class_definition =
             starknet_gateway_test_fixtures::class_definitions::CAIRO_0_11_SIERRA.to_vec();
 
-        let sierra_class = SierraHash(class2_hash.0);
-        let sierra_casm_hash = casm_hash_bytes!(b"non-existent");
-
         db_txn
-            .insert_cairo_class(class0_hash, &class0_definition)
+            .insert_cairo_class_definition(class0_hash, &class0_definition)
             .unwrap();
         db_txn
-            .insert_cairo_class(class1_hash, class1_definition)
+            .insert_cairo_class_definition(class1_hash, class1_definition)
             .unwrap();
         db_txn
-            .insert_sierra_class(
-                &sierra_class,
-                &sierra_class_definition,
-                &sierra_casm_hash,
-                &[],
-            )
+            .insert_sierra_class_definition(&sierra_class, &sierra_class_definition, &[])
             .unwrap();
         db_txn
-            .insert_cairo_class(class_hash_pending, &class0_definition)
+            .insert_cairo_class_definition(class_hash_pending, &class0_definition)
             .unwrap();
 
         // Update block 0
@@ -867,11 +862,12 @@ pub mod test_utils {
                 starknet_gateway_test_fixtures::class_definitions::CONTRACT_DEFINITION;
 
             for cairo in state_update_copy.declared_cairo_classes {
-                tx.insert_cairo_class(cairo, class_definition).unwrap();
+                tx.insert_cairo_class_definition(cairo, class_definition)
+                    .unwrap();
             }
 
-            for (sierra, casm) in state_update_copy.declared_sierra_classes {
-                tx.insert_sierra_class(&sierra, b"sierra def", &casm, b"casm def")
+            for (sierra, _casm) in state_update_copy.declared_sierra_classes {
+                tx.insert_sierra_class_definition(&sierra, b"sierra def", b"casm def")
                     .unwrap();
             }
 
@@ -1064,11 +1060,12 @@ pub mod test_utils {
                 starknet_gateway_test_fixtures::class_definitions::CONTRACT_DEFINITION;
 
             for cairo in state_update_copy.declared_cairo_classes {
-                tx.insert_cairo_class(cairo, class_definition).unwrap();
+                tx.insert_cairo_class_definition(cairo, class_definition)
+                    .unwrap();
             }
 
-            for (sierra, casm) in state_update_copy.declared_sierra_classes {
-                tx.insert_sierra_class(&sierra, b"sierra def", &casm, b"casm def")
+            for (sierra, _casm) in state_update_copy.declared_sierra_classes {
+                tx.insert_sierra_class_definition(&sierra, b"sierra def", b"casm def")
                     .unwrap();
             }
 
@@ -1420,18 +1417,20 @@ pub mod test_utils {
                 starknet_gateway_test_fixtures::class_definitions::CONTRACT_DEFINITION;
 
             for cairo in pre_latest_state_update.declared_cairo_classes {
-                tx.insert_cairo_class(cairo, class_definition).unwrap();
+                tx.insert_cairo_class_definition(cairo, class_definition)
+                    .unwrap();
             }
-            for (sierra, casm) in pre_latest_state_update.declared_sierra_classes {
-                tx.insert_sierra_class(&sierra, b"sierra def", &casm, b"casm def")
+            for (sierra, _casm) in pre_latest_state_update.declared_sierra_classes {
+                tx.insert_sierra_class_definition(&sierra, b"sierra def", b"casm def")
                     .unwrap();
             }
 
             for cairo in pre_confirmed_state_update_copy.declared_cairo_classes {
-                tx.insert_cairo_class(cairo, class_definition).unwrap();
+                tx.insert_cairo_class_definition(cairo, class_definition)
+                    .unwrap();
             }
-            for (sierra, casm) in pre_confirmed_state_update_copy.declared_sierra_classes {
-                tx.insert_sierra_class(&sierra, b"sierra def", &casm, b"casm def")
+            for (sierra, _casm) in pre_confirmed_state_update_copy.declared_sierra_classes {
+                tx.insert_sierra_class_definition(&sierra, b"sierra def", b"casm def")
                     .unwrap();
             }