+++ crates/rpc/src/method/estimate_message_fee.rs
@@ -304,13 +304,8 @@ mod tests {
 
             let class_hash =
                 class_hash!("0x032908a85d43275f8509ba5f2acae88811b293463a3521dc05ab06d534b40848");
-            tx.insert_sierra_class(
-                &SierraHash(class_hash.0),
-                sierra_json,
-                &casm_hash!("0x0564bc2cef7e8e8ded01da5999b2028ac5962669a12e12b33aee1b17b0332435"),
-                casm_json,
-            )
-            .expect("insert class");
+            tx.insert_sierra_class_definition(&SierraHash(class_hash.0), sierra_json, casm_json)
+                .expect("insert class");
 
             let block1_number = BlockNumber::GENESIS + 1;
             let block1_hash = BlockHash(felt!("0xb01"));